+++ src/onegov.people/onegov/people/tests/test_collections.py
@@ -16,16 +16,3 @@ def test_full_name(session):
 
     assert tom.title == 'Tom Chandler'
     assert rachel.title == 'Dr. Rachel Scott'
-
-
-def test_gravatar_url(session):
-    people = PersonCollection(session)
-
-    bill_gates = people.add(
-        first_name='Bill',
-        last_name='Gates',
-        email='billg@microsoft.com'
-    )
-
-    assert bill_gates.gravatar_url \
-        == 'https://www.gravatar.com/avatar/f9b0137cb32be76e5a171bff7ce98da7'