+++ src/onegov.people/HISTORY.rst
@@ -4,6 +4,9 @@ Changelog
 Unreleased
 ~~~~~~~~~~
 
+- Removes Gravatar support.
+  [href]
+
 0.0.1 (2015-07-03)
 ~~~~~~~~~~~~~~~~~~~